+++ changelog.md
@@ -23,6 +23,7 @@
     * `PotentialTransformerKind`: The construction kind of the potential transformer. (Enum)
     * `Sensor`: This class describes devices that transform a measured quantity into signals that can be presented at displays,
                 used in control or be recorded.
+* Added `PowerTransformer().getEnd(Terminal)`, which gets a `PowerTransformerEnd` by the `Terminal` it's connected to.
 
 ### Enhancements
 * None.